+++ packages/abap-deploy-config-inquirer/src/prompts/questions/config/package.ts
@@ -28,9 +28,9 @@ export function getPackagePrompts(options: AbapDeployConfigPromptOptions): Quest
     let morePackageResultsMsg = '';
     const isCli = !PromptState.isYUI;
 
-    const questions: Question[] = [
+    const questions: Question<AbapDeployConfigAnswersInternal>[] = [
         {
-            when: (): boolean => showPackageInputChoiceQuestion(),
+            when: (): boolean => showPackageInputChoiceQuestion(options.useAutocomplete),
             type: 'list',
             name: abapDeployConfigInternalPromptNames.packageInputChoice,
             message: t('prompts.config.package.packageInputChoice.message'),
@@ -71,10 +71,10 @@ export function getPackagePrompts(options: AbapDeployConfigPromptOptions): Quest
             },
             type: 'input',
             name: abapDeployConfigInternalPromptNames.packageCliExecution
-        },
+        } as InputQuestion<AbapDeployConfigAnswersInternal>,
         {
             when: (previousAnswers: AbapDeployConfigAnswersInternal): boolean =>
-                defaultOrShowManualPackageQuestion(isCli, previousAnswers.packageInputChoice),
+                defaultOrShowManualPackageQuestion(isCli, previousAnswers.packageInputChoice, options.useAutocomplete),
             type: 'input',
             name: abapDeployConfigInternalPromptNames.packageManual,
             message: t('prompts.config.package.packageManual.message'),
@@ -91,7 +91,7 @@ export function getPackagePrompts(options: AbapDeployConfigPromptOptions): Quest
         {
             when: (previousAnswers: AbapDeployConfigAnswersInternal): boolean =>
                 packageInputChoiceValid === true &&
-                defaultOrShowSearchPackageQuestion(isCli, previousAnswers.packageInputChoice),
+                defaultOrShowSearchPackageQuestion(isCli, previousAnswers.packageInputChoice, options.useAutocomplete),
             type: 'autocomplete',
             name: abapDeployConfigInternalPromptNames.packageAutocomplete,
             message: `${t('prompts.config.package.packageAutocomplete.message')}${