+++ packages/abap-deploy-config-inquirer/test/prompts/validators.test.ts
@@ -20,17 +20,21 @@ import {
     validateUrl
 } from '../../src/prompts/validators';
 import * as validatorUtils from '../../src/validator-utils';
-import {
-    AbapDeployConfigAnswersInternal,
-    ClientChoiceValue,
-    PackageInputChoices,
-    TargetSystemType,
-    TransportChoices
-} from '../../src/types';
+import { ClientChoiceValue, PackageInputChoices, TargetSystemType, TransportChoices } from '../../src/types';
 import * as utils from '../../src/utils';
 import { mockDestinations } from '../fixtures/destinations';
+import * as serviceProviderUtils from '../../src/service-provider-utils';
+
+jest.mock('../../src/service-provider-utils', () => ({
+    getTransportListFromService: jest.fn()
+}));
 
 describe('Test validators', () => {
+    const previousAnswers = {
+        url: 'https://mock.url.target1.com',
+        package: 'ZPACKAGE'
+    };
+
     beforeAll(async () => {
         await initI18n();
     });
@@ -73,7 +77,8 @@ describe('Test validators', () => {
                 client: '001',
                 destination: undefined,
                 isS4HC: false,
-                scp: false
+                scp: false,
+                targetSystem: 'https://mock.url.target1.com'
             });
             expect(result).toBe(true);
         });
@@ -185,7 +190,7 @@ describe('Test validators', () => {
 
     describe('validateCredentials', () => {
         it('should return error for no credentials', async () => {
-            expect(await validateCredentials('', {})).toBe(t('errors.requireCredentials'));
+            expect(await validateCredentials('', previousAnswers)).toBe(t('errors.requireCredentials'));
         });
 
         it('should return true for valid credentials', async () => {
@@ -193,7 +198,7 @@ describe('Test validators', () => {
                 transportConfig: {} as any,
                 transportConfigNeedsCreds: false
             });
-            expect(await validateCredentials('pass1', { username: 'user1' })).toBe(true);
+            expect(await validateCredentials('pass1', { ...previousAnswers, username: 'user1' })).toBe(true);
         });
 
         it('should return error message for invalid credentials', async () => {
@@ -201,7 +206,9 @@ describe('Test validators', () => {
                 transportConfig: {} as any,
                 transportConfigNeedsCreds: true
             });
-            expect(await validateCredentials('pass1', { username: 'user1' })).toBe(t('errors.incorrectCredentials'));
+            expect(await validateCredentials('pass1', { ...previousAnswers, username: 'user1' })).toBe(
+                t('errors.incorrectCredentials')
+            );
         });
     });
 
@@ -276,11 +283,17 @@ describe('Test validators', () => {
 
     describe('validatePackage', () => {
         it('should return error for invalid package input', async () => {
-            const result = await validatePackage('ZPACKAGE', {});
+            const getTransportListFromServiceSpy = jest.spyOn(serviceProviderUtils, 'getTransportListFromService');
+
+            const result = await validatePackage('zpackage', {
+                ...previousAnswers,
+                ui5AbapRepo: 'ZUI5REPO'
+            });
             expect(result).toBe(true);
+            expect(getTransportListFromServiceSpy).toBeCalledWith('ZPACKAGE', 'ZUI5REPO', {}, undefined);
         });
         it('should return error for invalid package input', async () => {
-            const result = await validatePackage(' ', {});
+            const result = await validatePackage(' ', previousAnswers);
             expect(result).toBe(t('warnings.providePackage'));
         });
     });
@@ -291,46 +304,51 @@ describe('Test validators', () => {
         });
 
         it('should return error for invalid package / ui5 abap repo name', async () => {
-            const previousAnswers: AbapDeployConfigAnswersInternal = {};
             let result = await validateTransportChoiceInput(TransportChoices.ListExistingChoice, previousAnswers);
             expect(result).toBe(t('errors.validators.transportListPreReqs'));
 
-            previousAnswers.packageManual = 'ZPACKAGE';
-            result = await validateTransportChoiceInput(TransportChoices.ListExistingChoice, previousAnswers);
+            result = await validateTransportChoiceInput(TransportChoices.ListExistingChoice, {
+                ...previousAnswers,
+                packageManual: 'ZPACKAGE'
+            });
             expect(result).toBe(t('errors.validators.transportListPreReqs'));
         });
 
         it('should return true for listing transport when transport request found for given config', async () => {
-            const previousAnswers: AbapDeployConfigAnswersInternal = {
-                packageManual: 'ZPACKAGE',
-                ui5AbapRepo: 'ZUI5REPO'
-            };
             jest.spyOn(validatorUtils, 'getTransportList').mockResolvedValueOnce([
                 { transportReqNumber: 'K123456', transportReqDescription: 'Mock transport request' }
             ]);
-            const result = await validateTransportChoiceInput(TransportChoices.ListExistingChoice, previousAnswers);
+            const result = await validateTransportChoiceInput(TransportChoices.ListExistingChoice, {
+                ...previousAnswers,
+                packageManual: 'ZPACKAGE',
+                ui5AbapRepo: 'ZUI5REPO'
+            });
             expect(result).toBe(true);
         });
 
         it('should return errors messages for listing transport when transport request empty or undefined', async () => {
-            const previousAnswers: AbapDeployConfigAnswersInternal = {
+            jest.spyOn(validatorUtils, 'getTransportList').mockResolvedValueOnce([]);
+
+            let result = await validateTransportChoiceInput(TransportChoices.ListExistingChoice, {
+                ...previousAnswers,
                 packageManual: 'ZPACKAGE',
                 ui5AbapRepo: 'ZUI5REPO'
-            };
-            jest.spyOn(validatorUtils, 'getTransportList').mockResolvedValueOnce([]);
-            let result = await validateTransportChoiceInput(TransportChoices.ListExistingChoice, previousAnswers);
+            });
             expect(result).toBe(t('warnings.noTransportReqs'));
 
             jest.spyOn(validatorUtils, 'getTransportList').mockResolvedValueOnce(undefined);
-            result = await validateTransportChoiceInput(TransportChoices.ListExistingChoice, previousAnswers);
+            result = await validateTransportChoiceInput(TransportChoices.ListExistingChoice, {
+                ...previousAnswers,
+                packageManual: 'ZPACKAGE',
+                ui5AbapRepo: 'ZUI5REPO'
+            });
             expect(result).toBe(t('warnings.noExistingTransportReqList'));
         });
 
         it('should return true if transport request is same as previous', async () => {
             const result = await validateTransportChoiceInput(
                 TransportChoices.CreateNewChoice,
-                {},
-
+                previousAnswers,
                 true,
                 TransportChoices.CreateNewChoice
             );
@@ -344,8 +362,7 @@ describe('Test validators', () => {
 
             const result = await validateTransportChoiceInput(
                 TransportChoices.CreateNewChoice,
-                {},
-
+                previousAnswers,
                 true,
                 undefined
             );
@@ -358,8 +375,7 @@ describe('Test validators', () => {
 
             const result = await validateTransportChoiceInput(
                 TransportChoices.CreateNewChoice,
-                {},
-
+                previousAnswers,
                 false,
                 undefined
             );
@@ -372,8 +388,7 @@ describe('Test validators', () => {
 
             const result = await validateTransportChoiceInput(
                 TransportChoices.CreateNewChoice,
-                {},
-
+                previousAnswers,
                 false,
                 undefined
             );
@@ -384,8 +399,7 @@ describe('Test validators', () => {
         it('should return error if creating a new transport request returns undefined', async () => {
             const result = await validateTransportChoiceInput(
                 TransportChoices.EnterManualChoice,
-                {},
-
+                previousAnswers,
                 false,
                 undefined
             );