+++ packages/fe-fpm-writer/src/building-block/index.ts
@@ -39,7 +39,7 @@ export function generateBuildingBlock<T extends BuildingBlock>(
     if (!fs) {
         fs = create(createStorage());
     }
-    validateBasePath(basePath, fs, ['sap.fe.core']);
+    validateBasePath(basePath, fs, ['sap.fe.templates', 'sap.fe.core']);
     if (!fs.exists(join(basePath, config.viewOrFragmentPath))) {
         throw new Error(`Invalid view path ${config.viewOrFragmentPath}.`);
     }