+++ packages/odata-service-inquirer/src/index.ts
@@ -2,13 +2,10 @@ import { type InquirerAdapter } from '@sap-ux/inquirer-common';
 import { type Logger } from '@sap-ux/logger';
 import { OdataVersion } from '@sap-ux/odata-service-writer';
 import { type ToolsSuiteTelemetryClient } from '@sap-ux/telemetry';
-import type { Question } from 'inquirer';
 import autocomplete from 'inquirer-autocomplete-prompt';
 import { ERROR_TYPE, ErrorHandler } from './error-handler/error-handler';
 import { initI18nOdataServiceInquirer } from './i18n';
 import { getQuestions } from './prompts';
-import type { AbapOnPremAnswers } from './prompts/datasources/sap-system/abap-on-prem/questions';
-import { getAbapOnPremSystemQuestions } from './prompts/datasources/sap-system/abap-on-prem/questions';
 import { newSystemChoiceValue } from './prompts/datasources/sap-system/new-system/questions';
 import LoggerHelper from './prompts/logger-helper';
 import {
@@ -19,8 +16,7 @@ import {
     type OdataServiceAnswers,
     type OdataServicePromptOptions,
     type OdataServiceQuestion,
-    type SapSystemType,
-    type SystemNamePromptOptions
+    type SapSystemType
 } from './types';
 import { PromptState, setTelemetryClient } from './utils';
 
@@ -89,30 +85,13 @@ async function prompt(
     return answers;
 }
 
-/**
- * Get the prompts for an abap on premise system. This can be used to create a new system connection.
- *
- * @param systemNamePromptOptions options for the system name prompt see {@link SystemNamePromptOptions}
- * @param logger a logger compatible with the {@link Logger} interface
- * @returns questions for creating a new abap on prem system connection
- */
-async function getAbapOnPremSystemPrompts(
-    systemNamePromptOptions: SystemNamePromptOptions,
-    logger?: Logger
-): Promise<Question<AbapOnPremAnswers>[]> {
-    if (logger) {
-        LoggerHelper.logger = logger;
-    }
-    // prompt texts must be loaded before the prompts are created, wait for the i18n bundle to be initialized
-    await initI18nOdataServiceInquirer();
-    return getAbapOnPremSystemQuestions(systemNamePromptOptions);
-}
-
 export {
     DatasourceType,
+    ERROR_TYPE,
+    ErrorHandler,
     OdataVersion,
-    getAbapOnPremSystemPrompts,
     getPrompts,
+    newSystemChoiceValue,
     prompt,
     promptNames,
     type CapRuntime,
@@ -120,9 +99,5 @@ export {
     type InquirerAdapter,
     type OdataServiceAnswers,
     type OdataServicePromptOptions,
-    type SapSystemType,
-    // These exports are to facilitate migration to open-ux-tools and will be removed in a future release
-    newSystemChoiceValue,
-    ERROR_TYPE,
-    ErrorHandler
+    type SapSystemType
 };