+++ packages/odata-service-inquirer/src/prompts/datasources/sap-system/new-system/types.ts
@@ -1,0 +1,17 @@
+import type { ODataVersion } from '@sap-ux/axios-extension';
+
+/**
+ * Sap System service answer
+ */
+export type ServiceAnswer = {
+    servicePath: string;
+    serviceODataVersion: ODataVersion;
+    toString: () => string;
+    serviceType?: string;
+};
+
+export const newSystemPromptNames = {
+    newSystemType: 'newSystemType',
+    newSystemUrl: 'newSystemUrl',
+    newSystemAuthType: 'newSystemAuthType'
+} as const;