+++ packages/abap-deploy-config-inquirer/CHANGELOG.md
@@ -1,5 +1,58 @@
 # @sap-ux/abap-deploy-config-inquirer
 
+## 0.0.9
+
+### Patch Changes
+
+-   99d7394: adds create command add deploy-config
+
+## 0.0.8
+
+### Patch Changes
+
+-   @sap-ux/axios-extension@1.16.5
+-   @sap-ux/fiori-generator-shared@0.3.18
+-   @sap-ux/system-access@0.5.10
+
+## 0.0.7
+
+### Patch Changes
+
+-   Updated dependencies [342e4be]
+    -   @sap-ux/guided-answers-helper@0.0.2
+
+## 0.0.6
+
+### Patch Changes
+
+-   Updated dependencies [1a99abc]
+    -   @sap-ux/ui5-config@0.24.1
+    -   @sap-ux/axios-extension@1.16.5
+    -   @sap-ux/fiori-generator-shared@0.3.17
+    -   @sap-ux/system-access@0.5.10
+
+## 0.0.5
+
+### Patch Changes
+
+-   fc58821: fix for local package verification
+
+## 0.0.4
+
+### Patch Changes
+
+-   Updated dependencies [d3dafeb]
+    -   @sap-ux/btp-utils@0.15.2
+    -   @sap-ux/axios-extension@1.16.5
+    -   @sap-ux/fiori-generator-shared@0.3.16
+    -   @sap-ux/system-access@0.5.10
+
+## 0.0.3
+
+### Patch Changes
+
+-   b158244: remove feature toggle usage
+
 ## 0.0.2
 
 ### Patch Changes