+++ packages/axios-extension/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/axios-extension
 
+## 1.16.5
+
+### Patch Changes
+
+-   Updated dependencies [d3dafeb]
+    -   @sap-ux/btp-utils@0.15.2
+
 ## 1.16.4
 
 ### Patch Changes