+++ packages/cards-editor-middleware/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/cards-editor-middleware
 
+## 0.4.21
+
+### Patch Changes
+
+-   Updated dependencies [df29368]
+    -   @sap-ux/project-access@1.27.0
+
+## 0.4.20
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.26.9
+
 ## 0.4.19
 
 ### Patch Changes