+++ packages/ui5-library-reference-inquirer/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/ui5-library-reference-inquirer
 
+## 0.3.27
+
+### Patch Changes
+
+-   Updated dependencies [df29368]
+    -   @sap-ux/project-access@1.27.0
+
+## 0.3.26
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.26.9
+
 ## 0.3.25
 
 ### Patch Changes